# docs/openapi.yaml
openapi: 3.0.1
info:
  title: TerminusDB API
  description: |
    API for the TerminusDB document graph database. Manage databases, documents, schemas, system, and queries.
    
    # Quick Start
    ```bash
    # List databases
    curl -u admin:root http://localhost:6363/api/
    
    # Create database
    curl -u admin:root -X POST http://localhost:6363/api/db/admin/mydb \
         -H "Content-Type: application/json" \
         -d '{"label": "My Database", "comment": "Description"}'
    
    # Insert unstructured JSON document into RDF database
    curl -u admin:root -X POST "http://localhost:6363/api/document/admin/mydb?raw_json=true&author=admin&message=new+document" \
         -H "Content-Type: application/json" \
         -d '{"name": "Alice", "age": 30}'

    # Insert schema for structured JSON documents
    curl -u admin:root -X POST "http://localhost:6363/api/document/admin/mydb?graph_type=schema&author=admin&message=new+schema" \
         -H "Content-Type: application/json" \
         -d '{"@type": "Class", "@id": "Person", "age": "xsd:integer", "name": "xsd:string"}'

    # Insert structured JSON-LD document
    curl -u admin:root -X POST "http://localhost:6363/api/document/admin/mydb?author=admin&message=new+document" \
         -H "Content-Type: application/json" \
         -d '{"@type": "Person", "age": 30, "name": "Alice"}'
    ```

...
  contact:
    email: support@dfrnt.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
...
servers:
...
security:
- basicAuth: []
...
paths:
...
  /duplicates/{path}:
    get:
      tags: [Search & Entity Resolution]
      operationId: duplicates
      summary: Scoped near-duplicate document groups (bounded).
      description: |
        Detects near-duplicate documents within a data product. For each
        indexed document in the set population, a nearest-neighbour query
        finds cross-document matches within `threshold`. Results are grouped
        at the document level (lower id first); a document is never grouped
        with itself.

        **Scope:** the `doc_type`/`doc_id` parameters define the set
        population. The `target_doc_type`/`target_doc_id` parameters, when
        present, define a second target population so that every group
        straddles set-to-target (cross-catalogue entity resolution). When
        omitted, the scan is within-set deduplication.

        `threshold` is on the [0, 1] cosine distance scale (0 = identical,
        0.5 = unrelated, 1 = opposite).

        **Identical documents:** when two documents contain literally identical
        content, the distance between them is 0; any difference yields a non-zero
        distance.

        **Authorization:** requires read access to the addressed data product.
        Denied callers receive 403 before any search is executed.
      parameters:
        - $ref: '#/components/parameters/search_path'
        - name: threshold
          in: query
          required: false
          description: |
            Maximum [0, 1] cosine distance for two documents to count as a
            near-duplicate (0 = identical, 0.5 = unrelated, 1 = opposite).
          schema:
            type: number
            format: float
            default: 0.0
        - $ref: '#/components/parameters/search_doc_type'
        - $ref: '#/components/parameters/search_doc_id'
        - name: target_doc_type
          in: query
          required: false
          description: |
            Restrict near-duplicate neighbours to these document types (the
            target population). Repeated key.
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
        - name: target_doc_id
          in: query
          required: false
          description: |
            Restrict near-duplicate neighbours to these document IDs (the
            target population). Repeated key.
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
        - $ref: '#/components/parameters/search_snippet'
        - $ref: '#/components/parameters/search_start'
        - $ref: '#/components/parameters/search_count'
      responses:
        200:
          description: |
            Array of near-duplicate groups, sorted nearest-first. Each group
            is a `group` array of members (lower id first) plus the [0, 1]
            cosine `distance` of the best matching chunks.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DuplicateGroups'
              examples:
                groups:
                  value:
                    - group:
                        - { id: "terminusdb:///data/Product/10" }
                        - { id: "terminusdb:///data/Product/57" }
                      distance: 0.06
                withSnippet:
                  summary: With snippet=true
                  value:
                    - group:
                        - { id: "terminusdb:///data/Product/10", snippet: "Sony 5 disc CD changer ..." }
                        - { id: "terminusdb:///data/Product/57", snippet: "Sony 5-disc CD player ..." }
                      distance: 0.04
        401:
          $ref: '#/components/responses/IncorrectAuthenticationError'
        403:
          $ref: '#/components/responses/ForbiddenError'
        404:
          description: No indexed ancestor exists for this branch.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchError'
        503:
          description: Search backend unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchError'
...
components:
  parameters:
...
    search_path:
      name: path
      in: path
      description: |
        Descriptor path identifying the data product branch to search. The path
        is resolved server-side to a domain (graphspec), HEAD commit, and ancestor
        window. Supports the standard TerminusDB path formats.
      required: true
      schema:
        type: string
      examples:
        default_branch:
          value: admin/mydb
          summary: Main branch of admin/mydb (autocompleted to admin/mydb/local/branch/main)
        explicit_branch:
          value: admin/mydb/local/branch/dev
          summary: Branch 'dev' of admin/mydb
    search_start:
      name: start
      in: query
      required: false
      description: Zero-based offset of the first result to return (pagination).
      schema:
        type: integer
        minimum: 0
        default: 0
    search_count:
      name: count
      in: query
      required: false
      description: Max results to return (pagination page size).
      schema:
        type: integer
        minimum: 1
        default: 50
...
    search_snippet:
      name: snippet
      in: query
      required: false
      description: |
        When true, include the matched chunk's text as `chunk.snippet` in each
        hit. Omitted/false keeps responses small.
      schema:
        type: boolean
        default: false
    search_doc_type:
      name: doc_type
      in: query
      required: false
      description: |
        Restrict to these document types. Repeated key
        (`?doc_type=People&doc_type=Species`). On POST, prefer the JSON array.
      schema:
        type: array
        items:
          type: string
      style: form
      explode: true
    search_doc_id:
      name: doc_id
      in: query
      required: false
      description: |
        Restrict to these full document IRIs. Repeated key. On POST, prefer
        the JSON array.
      schema:
        type: array
        items:
          type: string
      style: form
      explode: true
  schemas:
...
    SearchError:
      type: object
      additionalProperties: true
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable cause.
      example:
        error: "search backend unavailable"
...
    DuplicateGroups:
      type: array
      description: |
        Near-duplicate groups, sorted nearest-first. The `group` array is
        symmetric (member order is canonical: lower id first).
      items:
        type: object
        required: [group, distance]
        properties:
          group:
            type: array
            minItems: 2
            description: The member documents of this near-duplicate group.
            items:
              type: object
              required: [id]
              properties:
                id:
                  type: string
                  description: The member document id.
                snippet:
                  type: string
                  description: |
                    The member's matched chunk text. Present only when
                    `snippet=true` was requested.
          distance:
            type: number
            format: float
            description: |
              [0, 1] cosine distance of the group's best matching chunks
              (0 = identical, 0.5 = unrelated, 1 = opposite).
...
  responses:
    IncorrectAuthenticationError:
      description: Authentication information is missing or invalid
      content:
        application/json:
          schema:
            type: object
            properties:
              '@type':
                type: string
                example: api:ErrorResponse
              api:error:
                type: object
                properties:
                  '@type':
                    type: string
                    example: api:IncorrectAuthenticationError
              api:message:
                type: string
                example: Incorrect authentication information
              api:status:
                type: string
                example: api:failure
    ForbiddenError:
      description: User is authenticated but lacks permission for the requested action
      content:
        application/json:
          schema:
            type: object
            properties:
              api:status:
                type: string
                example: api:forbidden
              api:message:
                type: string
                example: Access to resource is not authorised with action read and auth User/anonymous
              auth:
                type: string
                example: User/anonymous
              action:
                type: string
                example: read
              scope:
                type: string
                example: Database/admin/test

    # ─── Search & Entity Resolution responses ────────────────────────────
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic